Add forwards approximation option, use +forward to enable
Adds a forwardApprox method to all PreOps that builds an automaton
overapproximating the language obtained by applying the operation to the
automaton constraints on the inputs.  E.g.

        z = concat(x, y)
        x \in Ax
        y \in Ay

will add the constraint z in Ax ++ Ay.  This is exact, but replaceall
is not regular in general and hence is approximated.  This hopefully
helps the search for a solution by restricting the solution space.