Fix another bug with word replacements
When matching e.g. abc, if an a is read instead of a b, then that a
could be the start of the new match. This was the old bug that was
fixed.

I didn't account for the fact that a could alternatively be the last
character read, and so there should be an accepting transition for that
case (you can't accept part way through a match, so existing transitions
didn't work).