Rework new forwards prop to return derived axioms
This meant changing a lot of the information collected.

First, remove the constraintStore -- it was only used for the full
exploration. The external solver will handle much of what it did now.
I.e. collect all new constraints and look for conflicts. (A separate
close rule is being implemented.)

Add to the information collected the original formula that it came from.
This allows an axiom to be created: argument constraint formulas +
application formula implies forwards propagated result formula.