Update SMTLIB test now we can solve it
failedProp should be unsat.

str.prefixof x (str.at x 1) can only be true if x is empty (in which
case (str.at x 1) is also empty. If x is not empty, either it is 2 or
more characters long -- so str.at x 1 returns a char -- or it is 1-character
long and str.at x 1 is empty. The prefixof is false in both cases.

So prefixof is only true when x is empty, but then it can't be different
from the truth of (= x ""). So no value of x will do.