Improve AutomatonTree equality checks with AutomataUtils
Rewrite the consistency check to also construct a witness word. This is
effectively a lazy product construction rather than the eager
construction that was being used before. The function can then be used
by the AutomatonTree to get distinguishing words for the nodes.

Since the witnessing word is also constructed, there's a tiny bit of
extra overhead to consistency checks. Hopefully this is a non-issue.